pyproject.toml
[project]
name = "chessgpt"
version = "2.0.0"
description = "A custom chess transformer that learns to predict and play chess moves"
requires-python = ">=3.11"
dependencies = [
    "torch>=2.4",
    "python-chess>=1.999",
    "tqdm>=4.66",
    "numpy>=2.0",
    "zstandard>=0.22",
]

[project.optional-dependencies]
dev = ["pytest", "ruff", "pre-commit"]
cloud = ["paramiko>=3.0", "runpod>=1.7", "python-dotenv>=1.0"]

[project.scripts]
chessgpt-download = "chessgpt.cli.download:main"
chessgpt-prepare = "chessgpt.cli.prepare:main"
chessgpt-train = "chessgpt.cli.train:main"
chessgpt-play = "chessgpt.cli.play:main"
chessgpt-eval = "chessgpt.cli.eval:main"
chessgpt-cloud = "chessgpt.cli.cloud:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/chessgpt"]

[tool.ruff]
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.ruff.lint.per-file-ignores]
"src/chessgpt/evaluation/puzzles.py" = ["E501"]  # chess notation strings

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
